[CSL-1871] Reorganize delegation modules in core
