+++ core/Pos/Aeson/Genesis.hs
@@ -13,6 +13,7 @@ import           Serokell.Aeson.Options (defaultOptions)
 import           Pos.Aeson.Core ()
 import           Pos.Aeson.Crypto ()
 import           Pos.Binary.Core.Address ()
+import           Pos.Core.Delegation.Types (ProxySKHeavy)
 import           Pos.Core.Genesis.Helpers (convertNonAvvmDataToBalances, recreateGenesisDelegation)
 import           Pos.Core.Genesis.Types (FakeAvvmOptions, GenesisAvvmBalances (..),
                                          GenesisDelegation, GenesisInitializer,
@@ -20,7 +21,7 @@ import           Pos.Core.Genesis.Types (FakeAvvmOptions, GenesisAvvmBalances (.
                                          GenesisVssCertificatesMap (..), GenesisWStakeholders (..),
                                          ProtocolConstants, TestnetBalanceOptions)
 import           Pos.Core.Ssc (VssCertificatesMap (..), validateVssCertificatesMap)
-import           Pos.Core.Types (ProxySKHeavy, StakeholderId)
+import           Pos.Core.Types (StakeholderId)
 import           Pos.Crypto (RedeemPublicKey, fromAvvmPk)
 import           Pos.Util.Util (eitherToFail)
