+++ core/Pos/Core/Genesis/Helpers.hs
@@ -17,8 +17,9 @@ import           Serokell.Util (allDistinct)
 import           Pos.Binary.Class (Bi)
 import           Pos.Core.Address (addressHash, decodeTextAddress)
 import           Pos.Core.Coin (unsafeAddCoin, unsafeIntegerToCoin)
+import           Pos.Core.Delegation.Types (ProxySKHeavy)
 import           Pos.Core.Genesis.Types (GenesisDelegation (..), GenesisNonAvvmBalances (..))
-import           Pos.Core.Types (Address, Coin, ProxySKHeavy, StakeholderId)
+import           Pos.Core.Types (Address, Coin, StakeholderId)
 import           Pos.Crypto.Signing (ProxySecretKey (..), isSelfSignedPsk)
 
 -- | Safe constructor of 'GenesisDelegation' from a list of PSKs.
