+++ core/Pos/Core/Types.hs
@@ -41,11 +41,6 @@ module Pos.Core.Types
        , HeaderHash
        , headerHashF
 
-       , ProxySigLight
-       , ProxySKLight
-       , ProxySigHeavy
-       , ProxySKHeavy
-
        , SharedSeed (..)
        , SlotLeaders
 
@@ -106,7 +101,7 @@ import           Pos.Core.Fee (TxFeePolicy)
 import           Pos.Core.Timestamp (TimeDiff (..), Timestamp (..))
 import           Pos.Crypto.Hashing (AbstractHash, Hash)
 import           Pos.Crypto.HD (HDAddressPayload)
-import           Pos.Crypto.Signing (ProxySecretKey, ProxySignature, PublicKey, RedeemPublicKey)
+import           Pos.Crypto.Signing (PublicKey, RedeemPublicKey)
 import           Pos.Data.Attributes (Attributes)
 
 ----------------------------------------------------------------------------
@@ -338,31 +333,6 @@ data BlockHeaderStub
 -- | Specialized formatter for 'HeaderHash'.
 headerHashF :: Format r (HeaderHash -> r)
 headerHashF = build
-
-----------------------------------------------------------------------------
--- Proxy signatures and delegation
-----------------------------------------------------------------------------
-
--- Notice: light delegation was removed as part of CSL-1856 and should
--- be reworked later. Though some parts of it are left to support
--- backward compatibility.
-
--- | Proxy signature, that holds a pair of epoch indices. Block is
--- valid if its epoch index is inside this range.
-type ProxySigLight a = ProxySignature (EpochIndex, EpochIndex) a
-
--- | Same alias for the proxy secret key (see 'ProxySigLight').
-type ProxySKLight = ProxySecretKey (EpochIndex, EpochIndex)
-
-
--- | Simple proxy signature without ttl/epoch index
--- constraints. 'EpochIndex' inside is needed for replay attack
--- prevention (it should match epoch of the block psk is announced
--- in).
-type ProxySigHeavy a = ProxySignature EpochIndex a
-
--- | Heavy delegation psk.
-type ProxySKHeavy = ProxySecretKey EpochIndex
 
 ----------------------------------------------------------------------------
 -- SSC. It means shared seed computation, btw
